requirements.txt
# Core dependencies for image processing pipeline
kafka-python-ng==2.2.2
minio==7.2.9
psycopg2-binary==2.9.9
watchdog==4.0.2
Pillow==10.4.0

# Optional - for local testing
python-dotenv==1.0.1
